move ptrace to linux module
