+++ elfcore/src/lib.rs
@@ -11,7 +11,6 @@ mod coredump;
 mod elf;
 mod error;
 mod linux;
-mod ptrace;
 
 pub use coredump::write_core_dump;
 pub use coredump::CoreDumpBuilder;
