+++ elfcore/src/linux/mod.rs
@@ -3,3 +3,4 @@
 
 pub mod memory;
 pub mod process;
+pub mod ptrace;
